remove migration utility
